pyproject.toml
[project]
name = "wagtail-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudinary>=1.42.1",
    "dj-database-url>=2.3.0",
    "django-browser-reload>=1.17.0",
    "django-cloudinary-storage>=0.3.0",
    "django-debug-toolbar>=5.0.1",
    "django-environ>=0.12.0",
    "django-extensions>=3.2.3",
    "django-tailwind>=3.8.0",
    "environ>=1.0",
    "gunicorn>=23.0.0",
    "pip>=24.3.1",
    "psycopg2-binary>=2.9.10",
    "wagtail>=6.3.2",
    "whitenoise>=6.8.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.1",
]
